+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Release (2025-XX-YY)
+- `stackitmarketplace`: [0.4.0](services/stackitmarketplace/CHANGELOG.md#v040-2025-04-11)
+  - **Feature:** Add new `InquiryContactSales`, `InquirySuggestProduct`, `PriceType`, `PricingOption` and `DeliveryMethod`
 - `stackitmarketplace`: [0.3.0](services/stackitmarketplace/CHANGELOG.md#v030-2025-04-04)
   - **Feature:** Add new `VendorProductId` attribute for subscription products